Moved dycore_state and geos_wrapper
