+++ fv3core/examples/standalone/runfile/dynamics.py
@@ -22,8 +22,10 @@ import pace.util as util
 from pace.dsl import StencilFactory
 from pace.dsl.dace.orchestration import DaceConfig
 from pace.fv3core import DynamicalCore, DynamicalCoreConfig
+
+# from pace.fv3core.initialization.dycore_state import DycoreState
+from pace.fv3core.dycore_state import DycoreState
 from pace.fv3core.initialization.baroclinic import init_baroclinic_state
-from pace.fv3core.initialization.dycore_state import DycoreState
 from pace.fv3core.testing import TranslateFVDynamics
 from pace.stencils.testing import dataset_to_dict
 from pace.stencils.testing.grid import Grid
