+++ fv3core/pace/fv3core/__init__.py
@@ -1,8 +1,12 @@
 from ._config import DynamicalCoreConfig
-from .initialization.dycore_state import DycoreState
-from .initialization.geos_wrapper import GeosDycoreWrapper
+
+# from .initialization.dycore_state import DycoreState
+from .dycore_state import DycoreState
 from .stencils.fv_dynamics import DynamicalCore
 from .stencils.fv_subgridz import DryConvectiveAdjustment
+
+# from .initialization.geos_wrapper import GeosDycoreWrapper
+from .wrappers.geos_wrapper import GeosDycoreWrapper
 
 
 __version__ = "0.2.0"
