+++ fv3core/pace/fv3core/initialization/baroclinic.py
@@ -9,7 +9,9 @@ import pace.fv3core.initialization.baroclinic_jablonowski_williamson as jablo_in
 import pace.util as fv3util
 import pace.util.constants as constants
 from pace.dsl.typing import Float
-from pace.fv3core.initialization.dycore_state import DycoreState
+
+# from pace.fv3core.initialization.dycore_state import DycoreState
+from pace.fv3core.dycore_state import DycoreState
 from pace.util.grid import GridData, lon_lat_midpoint
