+++ fv3core/pace/fv3core/initialization/init_utils.py
@@ -7,7 +7,9 @@ import numpy as np
 import pace.util as fv3util
 import pace.util.constants as constants
 from pace.dsl.typing import Float
-from pace.fv3core.initialization.dycore_state import DycoreState
+
+# from pace.fv3core.initialization.dycore_state import DycoreState
+from pace.fv3core.dycore_state import DycoreState
 from pace.util.grid import great_circle_distance_lon_lat, lon_lat_midpoint
 from pace.util.grid.gnomonic import get_lonlat_vect, get_unit_vector_direction
