+++ fv3core/pace/fv3core/stencils/fv_subgridz.py
@@ -14,7 +14,9 @@ import pace.dsl.gt4py_utils as utils
 import pace.util
 from pace.dsl.stencil import StencilFactory
 from pace.dsl.typing import Float, FloatField
-from pace.fv3core.initialization.dycore_state import DycoreState
+
+# from pace.fv3core.initialization.dycore_state import DycoreState
+from pace.fv3core.dycore_state import DycoreState
 from pace.fv3core.stencils.basic_operations import dim
 from pace.util import X_DIM, Y_DIM, Z_DIM
 from pace.util.constants import (
