+++ fv3core/pace/fv3core/testing/translate_fvdynamics.py
@@ -9,7 +9,9 @@ import pace.dsl.gt4py_utils as utils
 import pace.fv3core.stencils.fv_dynamics as fv_dynamics
 import pace.util
 from pace.fv3core._config import DynamicalCoreConfig
-from pace.fv3core.initialization.dycore_state import DycoreState
+
+# from pace.fv3core.initialization.dycore_state import DycoreState
+from pace.fv3core.dycore_state import DycoreState
 from pace.stencils.testing import ParallelTranslateBaseSlicing
 from pace.stencils.testing.translate import TranslateFortranData2Py
 from pace.util.grid import GridData
